fixed some bugs in position estimation - should be ready for some more outdoor tests now. Also added temporary suspension of tasks during waypoint communication to free the channel and speed it up, and fixed minor bug in waypoint receiving for overwriting existing waypoints while maintaining correct count

+++ Code/Library/sensing/gps_ublox.h
@@ -523,8 +523,8 @@ typedef struct
 	uint8_t num_sats;
 	uint16_t hdop;
 	
-	unsigned long timeLastMsg; //!< time reference in ms of microcontroller
-	unsigned long timegps; //!< time reference in ms of gps
+	uint32_t timeLastMsg; //!< time reference in ms of microcontroller
+	uint32_t timegps; //!< time reference in ms of gps
 	
 	unsigned char status;
